practice/08-CASE-二手车价格预测-P1/pyproject.toml
[project]
name = "used-car-price-prediction"
version = "1.0.0"
description = "二手车价格预测项目 - 机器学习实战"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.3.2",
    "numpy>=1.24,<2.0",
    "scikit-learn>=1.7.2",
    "xgboost>=3.0.5",
    "lightgbm>=4.6.0",
    "catboost>=1.2.8",
    "matplotlib>=3.10.6",
    "seaborn==0.12.2",
    "joblib>=1.5.2",
    "loguru>=0.7.3",
    "tqdm>=4.67.1",
    "python-dotenv>=1.1.1",
    "scipy>=1.4.1,<1.14",
    "mlxtend>=0.23.4",
    "bayesian-optimization>=2.0.2",
    "ipywidgets>=8.1.5",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "basedpyright>=1.32.1",
    "pandas-stubs>=2.3.2.250926",
]

[tool.basedpyright]
typeCheckingMode = "standard"
pythonVersion = "3.11"
pythonPlatform = "Darwin"